HTML file for live cryptocurrency  tracking
This project is an early implementation of a real-time cryptocurrency tracking website. It utilizes an API from a crypto company to fetch and display the latest values of various coins. The website is designed to provide users with up-to-date information on cryptocurrency prices, making it a useful tool for tracking market trends and making informed investment decisions.